@@ -13,20 +13,21 @@
   "main": "index.js",
   "dependencies": {
     "core-js": "^3.6.5",
-    "vue": "^2.6.11"
+    "vue": "^3.0.0"
   },
   "devDependencies": {
     "@vue/cli-plugin-babel": "~4.5.0",
     "@vue/cli-plugin-eslint": "~4.5.0",
     "@vue/cli-service": "~4.5.0",
+    "@vue/compiler-sfc": "^3.0.0",
     "autoprefixer": "^9.8.6",
     "babel-eslint": "^10.1.0",
     "eslint": "^6.8.0",
-    "eslint-plugin-vue": "^6.2.2",
+    "eslint-config-prettier": "^8.3.0",
+    "eslint-plugin-vue": "^7.0.0",
     "postcss": "^7.0.36",
     "prettier": "^2.3.2",
-    "tailwindcss": "npm:@tailwindcss/postcss7-compat@^2.2.7",
-    "vue-template-compiler": "^2.6.11"
+    "tailwindcss": "npm:@tailwindcss/postcss7-compat@^2.2.7"
   },
   "eslintConfig": {
     "root": true,
@@ -34,13 +35,16 @@
       "node": true
     },
     "extends": [
-      "plugin:vue/essential",
-      "eslint:recommended"
+      "plugin:vue/vue3-essential",
+      "prettier"
     ],
     "parserOptions": {
       "parser": "babel-eslint"
     },
-    "rules": {}
+    "rules": {
+      "vue/no-unused-vars": "error",
+      "vue/script-setup-uses-vars": "off"
+    }
   },
   "browserslist": [
     "> 1%",
